Add nw diagonal attacks.
This is nearly copy/paste from the prior commit.

It's probably completely fine to have these completely
parallel methods, but we should look out in the future
for some kind of better abstration.  It is pretty doubtful
that we will want to over-engineer this particular concern.

It's still way too early to get too finicky about code
quality.  We are very much in exploration mode.

https://www.epstudiossoftware.com/wp-content/uploads/2017/06/dry-300x208.png