EXPLORE: Start with one queen in lower left corner.
Let's start to actually attack the main problem at hand.
It's entirely plausible that at least one solution to the
N-queens problems simply starts with a queen in the lower
left corner.

Our assertions are already starting to get brittle, but this
may not actually be a major problem as long as we keep writing
fairly clean code.