Got 5 queens on the board.
In this commit we keep trying to add a queen on the board
until we run out of squares on the next rank.  The essential
part of the commit is that we introduce a "while" loop.

We are still in exploratory mode.

(Note that even for N=4 we only get a partial solution,
so we are gonna have to figure out some kind of backtracking
scheme.)