DEBUG: Get 7 queens on the board with absolutely brutal hack.
Here again we are in DEBUG/EXPLORATORY mode. If I am not smart
enough to solve the problem elegantly, I need to get a handle
on more brute force solutions. I am confident from prior
experiences solving this problem that I will quickly eliminate
this hack, but it is a great sanity check for now.