DEBUG: Temporarily disable attacks along the diagonals.
(In order to make sure that my approach to the N-queens
problem is at least plausibly viable, I want to go back
to the easier N=rooks problem.  This is a hacky way to
simulate N rooks instead of N queens.)

Note that the output to this version will misleadingly
show "Q" in squares still, even though we are treating
queens like they are rooks.  We will fix this soon.