MILESTONE: get all 92 solutions
This relatively simple commit again pays off some earlier
refactoring work.

The essence of this commit is that we simply don't stop
once we have a good solution.  Instead we keep yielding
good results.

A little bonus aspect of this commit is that we explicitly
assert that we got all 92 solutions. Since this problem
is well known, we can reasonably confident that getting 92
solutions means we have completely solved the problem!

Adding the very simple assertion here will allow us to clean
up the program with substantially less risk, which we will do
in upcoming commits.