Transpose how we place our queens.
We now try to place successive queens on the next FILE,
not the next RANK.

(FILE and RANK are chess terms for columns and rows.)

This is an example of "listening to the code". When we
were using the opposite approach of going rank by rank,
it led to awkard things like:

	return [(x, y) for y, x in enumerate(self.queens)]

The new way isn't necessarily much better, but part of the
purpose of this commit is to demonstrate that there are two
equally valid ways to attack a problem when symmetry is involved,
and as long as your fundamentals are solid, you can pretty easily
transpose between two different solutions.