CODE CLEANUP: Extract QueenMatingNet class.
The Board class was kinda long.  This seems like an easy
enough abstraction to pull out.

The name here is probably confusing if you don't play chess,
but the concept here is that the queens form a mating net
on a theoretical king.

I may have to rename this later. Hrmmmmmmmmm.