CODE CLEANUP: Remove most of the debugging/display code.
We don't need the debugging code any more, since we are 99.99%
confident that our core code works.  If we wanted to improve
the reliability of our program, we would want to think deeper
about adding some unit tests and assertions, as well as getting
code review.

As for the display code, it had some one-time value to show the
queens in action, but it was INCREDIBLY hacky, and, again, we
would want to think deeper about a better way to animate or
demonstrate our solution.